JWWEB-620: Fix JLink according to dev practices
